requirements.txt
# pip install -r requirements.txt
paho-mqtt==2.1.0